backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.15.0
numpy>=1.23.5
Pillow==9.3.0
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.26.0
tqdm==4.62.3
torch==2.0.1
transformers==4.30.2
